Remove artifact-specific pieces from base
